+++ CHANGELOG.md
@@ -3,13 +3,18 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
 ### Added
 
 - InitSegment.TweakSingleTrakLive changes an init segment to fit live streaming
+- File.UpdateSidx to update or add a top level sidx box for a fragmented file
+
+### Fixed
+
+- GetTrex() return value
 
 ## [0.42.0] - 2024-01-26